modularise models and export though index.js
