+++ src/models/airtable/index.js
@@ -1,0 +1,11 @@
+const {
+  queryApplicationsByFormula,
+  querySurveysByFormula,
+} = require("./queryByFormula")
+const queryById = require("./queryById")
+
+module.exports = {
+  queryApplicationsByFormula,
+  querySurveysByFormula,
+  queryById,
+}
