+++ src/models/email/index.js
@@ -1,0 +1,15 @@
+const sendCFNotification = require("./sendCFNotification")
+const sendClientNotification = require("./sendClientNotification")
+const sendClientInvitation = require("./sendClientInvitation")
+const sendClientInvitationReminder = require("./sendClientInvitationReminder")
+const sendFollowUpSurvey = require("./sendFollowUpSurvey")
+const sendClientSurveyNotification = require("./sendClientSurveyNotification")
+
+module.exports = {
+  sendCFNotification,
+  sendClientNotification,
+  sendClientInvitation,
+  sendClientInvitationReminder,
+  sendFollowUpSurvey,
+  sendClientSurveyNotification,
+}
