+++ src/models/github/index.js
@@ -1,0 +1,5 @@
+const addLabels = require("./addLabels")
+const createIssue = require("./createIssue")
+const updateIssue = require("./updateIssue")
+
+module.exports = { addLabels, createIssue, updateIssue }
